refactor: create e2e, ui test folders
