+++ app/src/androidTest/java/com/misw/vinilos/e2e/ArtistsE2ETest.kt
@@ -1,4 +1,4 @@
-package com.misw.vinilos
+package com.misw.vinilos.e2e
 
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.assertTextEquals
@@ -7,6 +7,7 @@ import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.LargeTest
+import com.misw.vinilos.MainActivity
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
