+++ app/src/androidTest/java/com/misw/vinilos/ui/AlbumDetailsScreenTest.kt
@@ -1,9 +1,10 @@
-package com.misw.vinilos
+package com.misw.vinilos.ui
 
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.misw.vinilos.DataFactory
 import com.misw.vinilos.ui.screens.albums.AlbumDetailsScreen
 import org.junit.Rule
 import org.junit.Test
