+++ app/src/androidTest/java/com/misw/vinilos/ui/CollectorsListScreenTest.kt
@@ -1,12 +1,13 @@
-package com.misw.vinilos
+package com.misw.vinilos.ui
 
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.misw.vinilos.ui.screens.artists.ArtistsListScreen
-import com.misw.vinilos.viewmodels.ArtistsViewModel
+import com.misw.vinilos.DataFactory
+import com.misw.vinilos.ui.screens.collectors.CollectorsListScreen
+import com.misw.vinilos.viewmodels.CollectorsViewModel
 import io.mockk.every
 import io.mockk.mockk
 import org.junit.Rule
@@ -14,7 +15,7 @@ import org.junit.Test
 import org.junit.runner.RunWith
 
 @RunWith(AndroidJUnit4::class)
-class ArtistsListScreenTest {
+class CollectorsListScreenTest {
 
     @get:Rule
     val composeTestRule = createComposeRule()
@@ -22,15 +23,15 @@ class ArtistsListScreenTest {
     @Test
     fun testLoadingState() {
         // Given
-        val mockViewModel = mockk<ArtistsViewModel>(relaxed = true)
+        val mockViewModel = mockk<CollectorsViewModel>(relaxed = true)
 
         // When
-        every { mockViewModel.artists } returns mutableStateOf(listOf())
+        every { mockViewModel.collectors } returns mutableStateOf(listOf())
         every { mockViewModel.isLoading } returns mutableStateOf(true)
         every { mockViewModel.hasError } returns mutableStateOf(false)
 
         composeTestRule.setContent {
-            ArtistsListScreen(viewModel = mockViewModel)
+            CollectorsListScreen(viewModel = mockViewModel)
         }
 
         // Then
@@ -40,15 +41,15 @@ class ArtistsListScreenTest {
     @Test
     fun testErrorState() {
         // Given
-        val mockViewModel = mockk<ArtistsViewModel>(relaxed = true)
+        val mockViewModel = mockk<CollectorsViewModel>(relaxed = true)
 
         // When
-        every { mockViewModel.artists } returns mutableStateOf(listOf())
+        every { mockViewModel.collectors } returns mutableStateOf(listOf())
         every { mockViewModel.isLoading } returns mutableStateOf(false)
         every { mockViewModel.hasError } returns mutableStateOf(true)
 
         composeTestRule.setContent {
-            ArtistsListScreen(viewModel = mockViewModel)
+            CollectorsListScreen(viewModel = mockViewModel)
         }
 
         // Then
@@ -56,21 +57,21 @@ class ArtistsListScreenTest {
     }
 
     @Test
-    fun testArtistListState() {
+    fun testCollectorListState() {
         // Given
-        val mockViewModel = mockk<ArtistsViewModel>(relaxed = true)
-        val artists = (1..5).map {DataFactory.createArtist()}
+        val mockViewModel = mockk<CollectorsViewModel>(relaxed = true)
+        val collectors = (1..5).map { DataFactory.createCollector() }
 
         // When
-        every { mockViewModel.artists } returns mutableStateOf(artists)
+        every { mockViewModel.collectors } returns mutableStateOf(collectors)
         every { mockViewModel.isLoading } returns mutableStateOf(false)
         every { mockViewModel.hasError } returns mutableStateOf(false)
 
         composeTestRule.setContent {
-            ArtistsListScreen(viewModel = mockViewModel)
+            CollectorsListScreen(viewModel = mockViewModel)
         }
 
         // Then
-        composeTestRule.onNodeWithTag("artistList").assertIsDisplayed()
+        composeTestRule.onNodeWithTag("collectorList").assertIsDisplayed()
     }
 }
